# faster Parse implementation