fix #3


```
Old Parse:
BenchmarkParseLong-8             1000000              1201 ns/op
BenchmarkParseShort-8            2000000               987 ns/op
BenchmarkParseErr-8             10000000               173 ns/op

New Parse:
BenchmarkParseLong-8            30000000                41.1 ns/op
BenchmarkParseShort-8           30000000                40.1 ns/op
BenchmarkParseErr-8             30000000                35.8 ns/op

```