@rultor 2019-01-27
> @rultor merge

@g4s8 OK, I'll try to merge now. You can check the progress of the merge [here](http://www.rultor.com/t/17552-457896242)