@rultor 2019-01-27
> @rultor merge

@g4s8 Done! FYI, the full log is [here](http://www.rultor.com/t/17552-457896242) (took me 2min)